Fetch partner payment terms and add them to generated invoice
We need custom marshalling and unmarshalling code for the payment terms,
as Odoo represents them as an array of `[ID, "Name"]` where ID is the
numerical ID and Name is the textual representation of the payment
terms.

The marshalling code has been modelled after the existing code for the
tax id.